Rename errors module to error and any references in the project - Update imports and docstrings for error in modules and tests - Update Sphinx documentation .rst files
