+++ src/cloudant/error.py
