+++ tests/unit/db/design_document_tests.py
@@ -27,7 +27,7 @@ import unittest
 from cloudant.document import Document 
 from cloudant.design_document import DesignDocument
 from cloudant.view import View, QueryIndexView
-from cloudant.errors import CloudantArgumentError, CloudantException
+from cloudant.error import CloudantArgumentError, CloudantException
 
 from .unit_t_db_base import UnitTestDbBase
